Finish infrastructure classes for long time matches

+++ DLC_for_WBFM/utils/training_data/tracklet_to_DLC.py
@@ -122,8 +122,8 @@ def convert_training_dataframe_to_scalar_format(df, min_length=10, scorer=None,
             this_brightness.append(segmentation_metadata.get_all_brightnesses(i_frame).iloc[i_local])
             this_volume.append(segmentation_metadata.get_all_volumes(i_frame).iloc[i_local])
         this_brightness = np.expand_dims(np.array(this_brightness), -1)
-        this_volume = np.expand_dims(np.array(this_volume), -1)
-        this_local_ind = np.expand_dims(np.array(this_local_ind), -1)
+        this_volume = np.expand_dims(np.array(this_volume, dtype=int), -1)
+        this_local_ind = np.expand_dims(np.array(this_local_ind, dtype=int), -1)
 
         # Combine all
         coords = np.hstack([zxy, confidence, this_local_ind, this_brightness, this_volume])
